backend/Similarity Check/requirements.txt
faiss-cpu
sentence-transformers
numpy